Add abstraction for addressing mode instead of Enum
We replaced the Enum implementation with a plain old class (subclass of object).
It could be done with DefaultEnumMeta, etc, but that's too complicated.

The code is simpler at the call site than before. The abstraction leaks slightly. This is
because False is returned upon failure setting the mode. And we then
raise an error with a message that assumes we know why it failed. So a change
to the AddressingMode class might require changes at the call sites.